requirements.txt
attrs==21.4.0
cffi==1.14.4
cryptography==3.3.2
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.20
PyMySQL==0.10.1
pyparsing==3.0.6
pytest==6.2.5
python-dotenv==0.15.0
six==1.15.0
toml==0.10.2
